+++ app/javascript/packs/workout_app.ts
@@ -3,7 +3,7 @@ import { createPinia } from 'pinia';
 import Modal from '@/components/Modal.vue';
 import { renderApp } from '@/shared/customized_vue';
 import { useElementPlus } from '@/shared/element_plus';
-import WorkoutApp from '@/workouts/Workout.vue';
+import WorkoutApp from '@/workout/Workout.vue';
 
 const app = renderApp(WorkoutApp);
 const pinia = createPinia();
